+++ src/app/catalogue/catalogue.component.spec.ts
