+++ src/app/components/cart-popover/cart-popover.component.ts
